+++ src/main/java/com/example/eat_together/domain/order/dto/response/OrderDetailResponseDto.java
@@ -1,15 +1,14 @@
-package com.example.eat_together.domain.order.dto;
+package com.example.eat_together.domain.order.dto.response;
 
 import com.example.eat_together.domain.order.entity.Order;
 import com.example.eat_together.domain.order.orderEnum.OrderStatus;
-import com.querydsl.core.annotations.QueryProjection;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
 @Getter
-public class OrderResponseDto {
-
+public class OrderDetailResponseDto {
     private final Long id;
 
     private final Long userId;
@@ -22,18 +21,28 @@ public class OrderResponseDto {
 
     private final double totalPrice;
 
+    private final List<OrderItemResponseDto> orderItems;
+
     private final LocalDateTime createdAt;
 
     private final LocalDateTime updatedAt;
 
-    @QueryProjection
-    public OrderResponseDto(Order order) {
+    public OrderDetailResponseDto(Order order) {
         this.id = order.getId();
         this.userId = order.getUser().getUserId();
         this.storeId = order.getStore().getStoreId();
         this.status = order.getStatus();
         this.deliveryFee = order.getStore().getDeliveryFee();
         this.totalPrice = order.getTotalPrice();
+        this.orderItems = order.getOrderItems().stream()
+                .map(item -> new OrderItemResponseDto(
+                        item.getMenu().getMenuId(),
+                        item.getMenu().getName(),
+                        item.getQuantity(),
+                        item.getPrice(),
+                        item.getPrice() * item.getQuantity()
+                ))
+                .toList();
         this.createdAt = order.getCreatedAt();
         this.updatedAt = order.getUpdatedAt();
     }
