+++ src/main/java/com/example/eat_together/domain/order/repository/OrderRepositoryCustom.java
@@ -1,6 +1,6 @@
 package com.example.eat_together.domain.order.repository;
 
-import com.example.eat_together.domain.order.dto.OrderResponseDto;
+import com.example.eat_together.domain.order.dto.response.OrderResponseDto;
 import com.example.eat_together.domain.order.entity.Order;
 import com.example.eat_together.domain.order.orderEnum.OrderStatus;
 import org.springframework.data.domain.Page;
