+++ src/main/java/com/example/eat_together/domain/order/service/OrderService.java
@@ -3,9 +3,9 @@ package com.example.eat_together.domain.order.service;
 import com.example.eat_together.domain.cart.entity.Cart;
 import com.example.eat_together.domain.cart.entity.CartItem;
 import com.example.eat_together.domain.cart.repository.CartRepository;
-import com.example.eat_together.domain.order.dto.OrderDetailResponseDto;
-import com.example.eat_together.domain.order.dto.OrderResponseDto;
-import com.example.eat_together.domain.order.dto.OrderStatusUpdateResponseDto;
+import com.example.eat_together.domain.order.dto.response.OrderDetailResponseDto;
+import com.example.eat_together.domain.order.dto.response.OrderResponseDto;
+import com.example.eat_together.domain.order.dto.response.OrderStatusUpdateResponseDto;
 import com.example.eat_together.domain.order.entity.Order;
 import com.example.eat_together.domain.order.entity.OrderItem;
 import com.example.eat_together.domain.order.orderEnum.OrderStatus;
